% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{bar_horizontal}
\alias{bar_horizontal}
\title{Horizontal bar chart}
\usage{
bar_horizontal(smry_data, x, y, div_by = NULL)
}
\arguments{
\item{smry_data}{summary data}

\item{x}{x variable}

\item{y}{y variable}

\item{div_by}{face wrapp according to an atribute}
}
\description{
Horizontal bar chart
}
