% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{bar_stacked_horizontal}
\alias{bar_stacked_horizontal}
\title{Stacked Bar Chart}
\usage{
bar_stacked_horizontal(
  smry_data,
  x,
  y,
  variable,
  categories = c("Erecto", "Semi-erecto", "Decumbente", "Postrado", "Semi-arrosetado",
    "Arrosetado"),
  color_categories = c("#BADF26", "#5CC665", "#22A385", "#297C8D", "#3B548A",
    "#472576")
)
}
\arguments{
\item{smry_data}{summary data}

\item{x}{group variable}

\item{y}{freq or count variable}

\item{variable}{variable or trait}

\item{categories}{vector of categories}

\item{color_categories}{color for each category}
}
\description{
Stacked Bar Chart
}
