% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_obs_data.R
\name{get_observation_data}
\alias{get_observation_data}
\title{Get obsevational data from WikiPapa Plataform}
\usage{
get_observation_data(
  url = "https://wikipapa.org/api/export/",
  call = "observations-data",
  idate = "2022-04-11"
)
}
\arguments{
\item{url}{character url path}

\item{call}{character api call name}

\item{idate}{character date format yyyy-mm-dd}
}
\description{
Observational data from the variety scouting
}
\author{
Omar Benites
}
