% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{get_summary_groupvar}
\alias{get_summary_groupvar}
\title{Get summary of variables from Wikipapa data}
\usage{
get_summary_groupvar(dfr, group, variable)
}
\arguments{
\item{dfr}{data data.frame or tibble}

\item{group}{character vector used to aggregate data by groups}

\item{variable}{chr variable}
}
\description{
Get summary of variables from Wikipapa data
}
