% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{get_summary_varcat}
\alias{get_summary_varcat}
\title{Summary of categorical variables}
\usage{
get_summary_varcat(dfr, catvar, total = FALSE)
}
\arguments{
\item{dfr}{data.frame data from WikiPapa platform}

\item{catvar}{character vector categorical variable}

\item{total}{logical If TRUE include total number of observations}
}
\description{
Summary of categorical variables
}
