% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/especies.R
\name{mutate_missing_catlabel}
\alias{mutate_missing_catlabel}
\title{Mutate missing category's label}
\usage{
mutate_missing_catlabel(dfr, variable, label = "No identificado")
}
\arguments{
\item{variable}{data.frame data from the catalogue}

\item{label}{character new label for the missing category}
}
\description{
Mutate missing category's label
}
