% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_box}
\alias{plot_box}
\title{Plot box}
\usage{
plot_box(variable, group = NULL, dots = c("no", "yes"), facet = "", dfr)
}
\arguments{
\item{variable}{variable}

\item{group}{group}

\item{dots}{character Whether yes, show dots}

\item{facet}{character facet off}

\item{dfr}{data}
}
\description{
Plot box
}
